requirements.txt
tensorflow>=2.15.0
streamlit>=1.30.0
pillow>=10.2.0
numpy>=1.26.0
tqdm>=4.66.0
matplotlib>=3.8.0
pandas>=2.2.0
scipy>=1.11.0
